+++ Demo/Demo/Tester.swift
@@ -260,7 +260,7 @@ class Tester {
 		}
 		
 		// cancel if need
-		//        task.cancel()
+		task.cancel()
 		
 		semaphore.wait()
 	}
