+++ Sources/UploadcareWidget/Utils.swift
@@ -1,0 +1,26 @@
+//
+//  Utils.swift
+//  
+//
+//  Created by Sergey Armodin on 21.04.2021.
+//  Copyright © 2021 Uploadcare, Inc. All rights reserved.`
+//
+
+import Foundation
+
+/// Debug log function with printing filename, method and line number
+///
+/// - Parameters:
+///   - messages: arguments
+///   - fullPath: filepath
+///   - line: line number
+///   - functionName: function/method name
+func DLog(_ messages: Any..., fullPath: String = #file, line: Int = #line, functionName: String = #function) {
+	#if DEBUG
+	let file = URL(fileURLWithPath: fullPath)
+	for message in messages {
+		let string = "\(file.pathComponents.last!):\(line) -> \(functionName): \(message)"
+		print(string)
+	}
+	#endif
+}
